grpc: move codegen into build.rs file
