+++ types/Cargo.toml
@@ -33,13 +33,13 @@ workspace-hack = { version = "0.1", path = "../workspace-hack" }
 
 [dev-dependencies]
 hex = "0.4.3"
-prost-build = "0.10.4"
 serde_test = "1.0.141"
-tonic-build = { version = "0.7.2", features = [ "prost", "transport" ] }
 criterion = "0.3.6"
 
 [build-dependencies]
 rustversion = "1.0.8"
+prost-build = "0.10.4"
+tonic-build = { version = "0.7.2", features = [ "prost", "transport" ] }
 
 [features]
 default = []
