+++ types/build.rs
@@ -1,11 +1,37 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-fn main() {
+use std::{env, path::PathBuf};
+
+type Result<T> = ::std::result::Result<T, Box<dyn std::error::Error + Send + Sync>>;
+
+fn main() -> Result<()> {
+    let out_dir = if env::var("DUMP_GENERATED_GRPC").is_ok() {
+        PathBuf::from("")
+    } else {
+        PathBuf::from(env::var("OUT_DIR")?)
+    };
+
+    let proto_files = &["proto/narwhal.proto"];
+    let dirs = &["proto"];
+
+    // Use `Bytes` instead of `Vec<u8>` for bytes fields
+    let mut config = prost_build::Config::new();
+    config.bytes(&["."]);
+
+    tonic_build::configure()
+        .out_dir(out_dir)
+        .compile_with_config(config, proto_files, dirs)?;
+
     println!("cargo:rerun-if-changed=build.rs");
+    println!("cargo:rerun-if-changed=proto");
+    println!("cargo:rerun-if-env-changed=DUMP_GENERATED_GRPC");
+
     nightly();
     beta();
     stable();
+
+    Ok(())
 }
 
 #[rustversion::nightly]
