+++ types/src/proto.rs
@@ -1,10 +1,9 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-#[cfg_attr(beta, allow(clippy::derive_partial_eq_without_eq))]
-#[path = "generated/narwhal.rs"]
-#[rustfmt::skip]
-mod narwhal;
+mod narwhal {
+    tonic::include_proto!("narwhal");
+}
 
 use std::{array::TryFromSliceError, ops::Deref};
